+++ src/FileAttachment/Contracts/DraftAttachmentHandler.php
@@ -2,13 +2,24 @@
 
 namespace CbtechLtd\Fastlane\FileAttachment\Contracts;
 
-use CbtechLtd\Fastlane\FileAttachment\DraftAttachment;
+use CbtechLtd\Fastlane\FileAttachment\AttachmentValue;
 use CbtechLtd\Fastlane\Http\Requests\EntryAttachmentStoreRequest;
 use CbtechLtd\Fastlane\Support\Schema\Fields\Contracts\HasAttachments;
+use Illuminate\Support\Collection;
 
 interface DraftAttachmentHandler
 {
-    public function __construct(HasAttachments $field);
+    /**
+     * @param string     $draftId
+     * @param array|null $files
+     * @return Collection<AttachmentValue>
+     */
+    public function findDrafts(string $draftId, ?array $files = null): Collection;
 
-    public function handle(EntryAttachmentStoreRequest $request): DraftAttachment;
+    /**
+     * @param EntryAttachmentStoreRequest $request
+     * @param HasAttachments              $field
+     * @return AttachmentValue
+     */
+    public function write(EntryAttachmentStoreRequest $request, HasAttachments $field): AttachmentValue;
 }